Move function out of model file and into a better home
